Split app store selection helpers to satisfy structural lint
